SearchPage takes a rowExtras render prop; ScoreBreakdown moves to diagnostic chunk
Previous commit moved /search out of the main bundle but ScoreBreakdown
still sat inside SearchPage, so it shipped to anyone who navigated to
/search even though only /search/test ever rendered it.

Now SearchPage takes a rowExtras render prop:
  rowExtras?: (item: MergedSearchItem) => ReactNode

PhraseResultRow and EntityResultRow each accept an opaque `extras`
node and slot it under the type badge in the right column. The score
component itself moved to src/routes/_user/-score-breakdown.tsx and is
imported only by search.test.lazy.tsx.

Vite splits accordingly:
- /search loads SearchPage chunk only (20.9 KB, was 22.4 KB).
- /search/test loads SearchPage + ScoreBreakdown bundled into its
  route chunk (2.0 KB, was 926 B). Net same total bytes; the
  diagnostic-only code now lives in the diagnostic-only chunk.

Drops the `diagnostic={boolean}` prop pattern in favor of explicit
component injection.
